WIP: Move CSS component into components folder
Changes:
* Introduce png/jpg/gif image loader
* All style components moved into component folders
* Deleted old unused LESS artifact
* public folder now added to .gitignore
* Output page title

Issues:
* componentDidMount not running - is JS being run on server?
