+++ .gitignore
@@ -1,3 +1,4 @@
 npm-debug.log
 node_modules/
 db_collection/
+public/dist/*
