+++ app.js
@@ -1,3 +1,5 @@
 // app.js
-require('babel/register')
+require('babel-core/register')
+// Avoid processing of CSS on the server before continuing
+require.extensions['.css'] = function () { return null }
 require('./libs/server/app-server.js')
