+++ libs/client/api.js
@@ -1,4 +1,27 @@
-var Api = require( './libs/Api.js')
+var fetch = require( 'isomorphic-fetch' );
+function Api() {}
+
+Api.prototype = {
+  getPage: function ( title, lang ) {
+    lang = lang || 'en';
+    return fetch( '/api/' + lang + '/' + title ).then( function ( resp ) {
+      return resp.json();
+    } );
+  },
+  getTrending: function ( filter ) {
+    var url = '/api/trending/';
+    if ( filter ) {
+      url += filter;
+    }
+    return fetch( url ).then( function ( resp ) {
+      return resp.json();
+    } ).then( function ( json ) {
+      return json.results;
+    } );
+  }
+};
+
+export default Api;
 let api = new Api();
 
 export default api;
