--- libs/client/libs/Api.js
@@ -1,24 +1,0 @@
-var fetch = require( 'isomorphic-fetch' );
-function Api() {}
-
-Api.prototype = {
-  getPage: function ( title, lang ) {
-    lang = lang || 'en';
-    return fetch( '/api/' + lang + '/' + title ).then( function ( resp ) {
-      return resp.json();
-    } );
-  },
-  getTrending: function ( filter ) {
-    var url = '/api/trending/';
-    if ( filter ) {
-      url += filter;
-    }
-    return fetch( url ).then( function ( resp ) {
-      return resp.json();
-    } ).then( function ( json ) {
-      return json.results;
-    } );
-  }
-};
-
-export default Api;
