+++ libs/client/main.css
@@ -1,1 +1,2 @@
+#app { padding-top: 12px; }
 body { background: #f0f0f0; }
