+++ libs/client/views/components/IntermediateState/index.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react'
+import './styles.css'
 
 class IntermediateState extends Component {
   render(){
