+++ libs/client/views/components/IntermediateState/styles.css
@@ -8,5 +8,5 @@
   padding-left: 0.4em;
 	box-shadow: inset 0 0 0 0.1em #ffffff;
 	transition: border 200ms cubic-bezier(0.39,0.575,0.565,1), box-shadow 200ms cubic-bezier(0.39,0.575,0.565,1);
-	background-image: url(pending.gif)
+	background-image: url("pending.gif")
 }
