+++ libs/client/views/components/Topic/index.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react'
+import './styles.css'
 
 class Topic extends Component {
   render(){
