+++ libs/server/views/index.html
@@ -5,16 +5,9 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <meta name="description" content="">
   <meta name="author" content="">
-  <title>{{ site.title }}{{# page }} | {{ page.title }}{{/ page }}</title>
+  <title>Weekapedia</title>
   <!-- Bootstrap Core CSS -->
-  <link href="/dist/mediawiki.bootstrap.css" rel="stylesheet">
   <link href="/dist/style.css" rel="stylesheet">
-  <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
-  <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-  <!--[if lt IE 9]>
-    <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>
-    <script src="https://oss.maxcdn.com/libs/respond.js/1.4.2/respond.min.js"></script>
-  <![endif]-->
 </head>
 <body class="hidden">
   <div id="app">{{{ reactMarkup }}}</div>
