+++ package.json
@@ -4,36 +4,45 @@
   "description": "",
   "main": "app-server.js",
   "dependencies": {
-    "autoprefixer-loader": "^3.2.0",
-    "babel": "^5.8.29",
-    "babel-core": "^5.8.32",
     "cosmicjs": "^2.0.0",
-    "css-loader": "^0.23.1",
     "events": "^1.1.0",
     "express": "^4.13.3",
-    "extract-text-webpack-plugin": "^1.0.1",
     "flux": "^2.1.1",
-    "history": "^1.14.0",
     "hogan-express": "^0.5.2",
     "isomorphic-fetch": "^2.2.1",
     "lodash": "^3.10.1",
-    "react": "^0.14.1",
-    "react-dom": "^0.14.1",
-    "react-router": "^1.0.1",
     "webpack": "^1.12.2"
   },
   "scripts": {
-    "development": "cp libs/server/views/index.html public/index.html && NODE_ENV=development webpack && webpack-dev-server --content-base public/ --hot --inline --devtool inline-source-map --history-api-fallback",
-    "production": "rm -rf public/index.html && NODE_ENV=production webpack -p && NODE_ENV=production node app.js",
+    "node-server-dev": "NODE_ENV=development node app.js",
+    "development": "npm run webpack-dev & npm run node-server-dev",
+    "webpack-dev": "cp libs/server/views/index.html public/index.html && NODE_ENV=development webpack && webpack-dev-server --content-base public/ --hot --inline --devtool inline-source-map --history-api-fallback",
+    "production": "rm -rf public/index.html && NODE_ENV=production webpack -p ",
     "start": "npm run production"
   },
   "author": "",
   "license": "ISC",
+  "babel": {
+    "presets": [
+      "es2015",
+      "react"
+    ]
+  },
   "devDependencies": {
+    "autoprefixer-loader": "^3.2.0",
     "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-react": "^6.11.1",
+    "css-loader": "^0.23.1",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "file-loader": "^0.9.0",
     "node-cache": "^3.2.1",
+    "react": "^15.2.0",
+    "react-dom": "^15.2.0",
     "react-hot-loader": "^1.3.0",
+    "react-router": "^2.5.2",
     "style-loader": "^0.13.1",
+    "url-loader": "^0.5.7",
     "webpack-dev-server": "^1.12.1"
   }
 }
